Restructure projects, closes #9
