+++ blottertrax/services/soundcloud.py
@@ -1,7 +1,7 @@
 import requests
 
-from config import Config
-from serviceresult import ServiceResult
+from blottertrax.config import Config
+from blottertrax.value_objects.serviceresult import ServiceResult
 
 
 class Soundcloud:
